Created a utility function that formats the duedate before displaying in the table.
